+++ src/Components/FlashContainer.vue
@@ -1,15 +1,23 @@
-<script setup>
+<script setup lang="ts">
 import FlashMessage from "./FlashMessage.vue";
-import {flash} from "../Utils/flash.ts";
+import {flash, FlashVisuals, defaultFlashVisuals} from "../Utils/flash";
+import {VisualParam, VisualParamDef} from "../Utils/VisualParam.js";
 import {ref} from "vue";
+
+const props = defineProps<{
+	visual?: VisualParamDef<FlashVisuals>;
+}>();
+
+const visual = new VisualParam<FlashVisuals>();
+if (props.visual) visual.addParamDef(props.visual);
+visual.addParamDef(defaultFlashVisuals);
+const vis = visual.getAll();
 
 const messages = ref([]);
 let lastId = 0;
 const add = (type, message) => {
 	lastId++;
-	messages.value.push({
-		message, type, id: lastId
-	});
+	messages.value.push({ message, type, id: lastId });
 }
 const remove = (id) => {
 	messages.value = messages.value.filter(item => item.id !== id);
@@ -18,8 +26,16 @@ flash.registerAdd((type, message) => add(type, message));
 </script>
 
 <template>
-<div :class="flash.containerCssClasses()">
-	<FlashMessage v-for="message in messages" :id="message.id" :message="message.message" :type="message.type" @remove="remove" :key="message.id" />
+<div :class="vis.containerClass">
+	<FlashMessage
+		v-for="message in messages"
+		:id="message.id"
+		:message="message.message"
+		:type="message.type"
+		@remove="remove"
+		:key="message.id"
+		:visuals="props.visual"
+	/>
 </div>
 </template>